refactor: cartesian.py
